Extracted editor warnings to a custom hook
